workers/video-renderer/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
boto3==1.29.7
python-multipart==0.0.6
ffmpeg-python==0.2.0
opencv-python==4.8.1.78
numpy>=1.24.3
Pillow>=10.0.0
pysrt>=1.1.2

